quic: use shared function for allow unauthorized check
