+++ lib/internal/quic/util.js
@@ -19,7 +19,10 @@ const {
   isIP,
 } = require('internal/net');
 
-const { getOptionValue } = require('internal/options');
+const {
+  getOptionValue,
+  getAllowUnauthorized,
+} = require('internal/options');
 
 const {
   sessionConfig,
@@ -65,31 +68,12 @@ const {
   }
 } = internalBinding('quic');
 
-let warnOnAllowUnauthorized = true;
 let dns;
 
 function lazyDNS() {
   if (!dns)
     dns = require('dns');
   return dns;
-}
-
-function GetMaxHeaderLength() {
-  getOptionValue('--max-http-header-size') | 0;
-}
-
-function getAllowUnauthorized() {
-  const allowUnauthorized = process.env.NODE_TLS_REJECT_UNAUTHORIZED === '0';
-
-  if (allowUnauthorized && warnOnAllowUnauthorized) {
-    warnOnAllowUnauthorized = false;
-    process.emitWarning(
-      'Setting the NODE_TLS_REJECT_UNAUTHORIZED ' +
-      'environment variable to \'0\' makes TLS connections ' +
-      'and HTTPS requests insecure by disabling ' +
-      'certificate verification.');
-  }
-  return allowUnauthorized;
 }
 
 function getSocketType(type = 'udp4') {
@@ -196,7 +180,7 @@ function validateTransportParams(params) {
       maxHeaderListSize,
       maxPushes,
       maxHeaderPairs,
-      maxHeaderLength = GetMaxHeaderLength(),
+      maxHeaderLength = getOptionValue('--max-http-header-size'),
     },
   } = { h3: {}, ...params };
   validateNumberInBoundedRange(
